app/requirements.txt
# Minimal requirements for inference only
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.0.0
tqdm>=4.64.0
